Propagate/document TooManyDescriptors error return
This can happen on any component call. On component calls which used `?`
or a similar method for mapping error codes, it was already possible for
it to be returned, but it was not documented. On component calls which
used heavily custom error mapping, it may have been squashed into
`BadComponent`, which was not proper handling.